+++ design-system/textfield/examples/00-html-props.tsx
@@ -10,9 +10,9 @@ export default function HtmlPropsExample() {
   return (
     <div>
       <Label htmlFor="password">Password text field</Label>
-      <Textfield name="password" type="password" />
+      <Textfield name="password" type="password" id="password" />
       <Label htmlFor="number">Number field (with min/max values)</Label>
-      <Textfield name="password" type="number" max={5} min={0} />
+      <Textfield name="number" type="number" id="number" max={5} min={0} />
     </div>
   );
 }