+++ design-system/textfield/examples/constellation/text-field-default.tsx
@@ -3,5 +3,5 @@ import React from 'react';
 import Textfield from '../../src';
 
 export default function TextFieldDefaultExample() {
-  return <Textfield name="basic" />;
+  return <Textfield name="basic" aria-label="default text field" />;
 }